smart_factory/requirements.txt
paho-mqtt>=2.0.0
numpy>=1.24.0
streamlit>=1.25.0
